refactor: member Controller 패키지 변경
