+++ src/main/java/dandi/dandi/member/adapter/in/web/MemberBlockControllerDocs.java
@@ -1,4 +1,4 @@
-package dandi.dandi.member.web;
+package dandi.dandi.member.adapter.in.web;
 
 import dandi.dandi.member.application.port.in.MemberBlockCommand;
 import io.swagger.v3.oas.annotations.Operation;
