+++ src/main/java/dandi/dandi/member/adapter/in/web/dto/in/NicknameUpdateRequest.java
@@ -1,7 +1,6 @@
-package dandi.dandi.member.web.dto.in;
+package dandi.dandi.member.adapter.in.web.dto.in;
 
 import dandi.dandi.member.application.port.in.NicknameUpdateCommand;
-import io.swagger.v3.oas.annotations.media.Schema;
 
 public class NicknameUpdateRequest {
