+++ src/test/java/dandi/dandi/member/acceptance/MemberAcceptanceTest.java
@@ -28,12 +28,12 @@ import dandi.dandi.auth.web.in.LoginRequest;
 import dandi.dandi.comment.application.port.in.CommentResponse;
 import dandi.dandi.comment.application.port.in.CommentResponses;
 import dandi.dandi.common.AcceptanceTest;
+import dandi.dandi.member.adapter.in.web.dto.in.LocationUpdateRequest;
+import dandi.dandi.member.adapter.in.web.dto.in.NicknameUpdateRequest;
 import dandi.dandi.member.application.port.in.MemberBlockCommand;
 import dandi.dandi.member.application.port.in.MemberInfoResponse;
 import dandi.dandi.member.application.port.in.NicknameDuplicationCheckResponse;
 import dandi.dandi.member.application.port.in.ProfileImageUpdateResponse;
-import dandi.dandi.member.web.dto.in.LocationUpdateRequest;
-import dandi.dandi.member.web.dto.in.NicknameUpdateRequest;
 import dandi.dandi.post.application.port.in.FeedResponse;
 import dandi.dandi.post.application.port.in.PostResponse;
 import io.restassured.response.ExtractableResponse;
